requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
mysql-connector==2.2.9
pygame==2.5.2
requests==2.31.0
urllib3==2.1.0
yarl==1.9.4
